requirements.txt
numpy==1.24.4
pandas==1.5.3
psycopg2-binary==2.9.9
requests==2.32.3
SQLAlchemy==1.4.53
sqlalchemy-redshift==0.8.14 
apache-airflow==2.10.0
python-dotenv
awswrangler
